+++ src/lib/hashtags.ts
@@ -1,8 +1,6 @@
 import { findAndReplace, Replace } from "mdast-util-find-and-replace"
 import { Plugin } from "unified"
 
-import { allPosts } from "contentlayer/generated"
-
 /**
  * Plugin to autolink references for hashtags.
  */
@@ -17,17 +15,6 @@ export const remarkHashtags: Plugin = () => {
       }
     )
   }
-}
-
-export const allTags = (): string[] => {
-  const tags = new Set<string>()
-  for (const post of allPosts) {
-    for (const tag of post.tags) {
-      tags.add(tag)
-    }
-  }
-
-  return Array.from(tags).sort((a, b) => a.localeCompare(b))
 }
 
 const replaceHashtag: Replace = (
